+++ mongoengine/base.py
@@ -470,6 +470,8 @@ class ObjectIdField(BaseField):
     """
 
     def to_python(self, value):
+        if not isinstance(value, ObjectId):
+            value = ObjectId(value)
         return value
 
     def to_mongo(self, value):