+++ tests/test_all_warnings.py
@@ -37,6 +37,28 @@ class TestWarnings(unittest.TestCase):
         self.assertEqual(FutureWarning, warning["category"])
         self.assertTrue("InheritedClass" in str(warning["message"]))
 
+    def test_dbref_reference_field_future_warning(self):
+
+        class Person(Document):
+            name = StringField()
+            parent = ReferenceField('self')
+
+        Person.drop_collection()
+
+        p1 = Person()
+        p1.parent = None
+        p1.save()
+
+        p2 = Person(name="Wilson Jr")
+        p2.parent = p1
+        p2.save(cascade=False)
+
+        self.assertEqual(len(self.warning_list), 1)
+        warning = self.warning_list[0]
+        self.assertEqual(FutureWarning, warning["category"])
+        self.assertTrue("ReferenceFields will default to using ObjectId"
+                        in str(warning["message"]))
+
     def test_document_save_cascade_future_warning(self):
 
         class Person(Document):